Move dbs and seeder
