+++ src/tribler/core/conftest.py
@@ -7,17 +7,12 @@ from unittest.mock import MagicMock
 
 import pytest
 from ipv8.keyvault.crypto import default_eccrypto
-from ipv8.keyvault.private.libnaclkey import LibNaCLSK
 
-from tribler.core.components.knowledge.db.knowledge_db import KnowledgeDatabase
 from tribler.core.components.libtorrent.download_manager.download import Download
 from tribler.core.components.libtorrent.download_manager.download_config import DownloadConfig
 from tribler.core.components.libtorrent.download_manager.download_manager import DownloadManager
 from tribler.core.components.libtorrent.settings import LibtorrentSettings
-from tribler.core.components.metadata_store.db.store import MetadataStore
-from tribler.core.tests.tools.common import TESTS_DATA_DIR
 from tribler.core.utilities.network_utils import default_network_utils
-from tribler.core.utilities.simpledefs import DownloadStatus
 from tribler.core.utilities.unicode import hexlify
 
 # Enable origin tracking for coroutine objects in the current thread, so when a test does not handle
@@ -44,52 +39,6 @@ def pytest_runtest_protocol(item, log=True, nextitem=None):  # pylint: disable=u
 @pytest.fixture
 def free_port():
     return default_network_utils.get_random_free_port(start=1024, stop=50000)
-
-
-@pytest.fixture
-async def video_seeder(tmp_path_factory, video_tdef):
-    config = LibtorrentSettings()
-    config.dht = False
-    config.upnp = False
-    config.natpmp = False
-    config.lsd = False
-    seeder_state_dir = tmp_path_factory.mktemp('video_seeder_state_dir')
-    dlmgr = DownloadManager(
-        config=config,
-        state_dir=seeder_state_dir,
-        notifier=MagicMock(),
-        peer_mid=b"0000")
-    dlmgr.metadata_tmpdir = tmp_path_factory.mktemp('metadata_tmpdir')
-    dlmgr.initialize()
-    dscfg_seed = DownloadConfig()
-    dscfg_seed.set_dest_dir(TESTS_DATA_DIR)
-    upload = dlmgr.start_download(tdef=video_tdef, config=dscfg_seed)
-    await upload.wait_for_status(DownloadStatus.SEEDING)
-    yield dlmgr
-    await dlmgr.shutdown()
-
-
-TEST_PERSONAL_KEY = LibNaCLSK(
-    b'4c69624e61434c534b3af56022aa5d556c07aeed704ee98df7dca580f'
-    b'522e1405663f0d36508d2189cb8991af2dd27b34bc18b4d24869e2c4f2cfdb164a78ea6e687daf7a21640d62b1b'[10:]
-)
-
-
-@pytest.fixture
-def metadata_store(tmp_path):
-    mds = MetadataStore(db_filename=tmp_path / 'test.db',
-                        channels_dir=tmp_path / 'channels',
-                        my_key=TEST_PERSONAL_KEY,
-                        disable_sync=True)
-    yield mds
-    mds.shutdown()
-
-
-@pytest.fixture
-def knowledge_db():
-    db = KnowledgeDatabase()
-    yield db
-    db.shutdown()
 
 
 @pytest.fixture
